refactor: replace handler with rpcHandler and reorganize router structure
