+++ libs/orpc/src/index.ts
@@ -1,4 +1,4 @@
 export { client } from "./client";
 export { type Context, createContext } from "./context";
-export { handler } from "./handler";
-export { appRouter } from "./routers/app/router";
+export { appRouter } from "./routers/app.router";
+export { rpcHandler } from "./rpcHandler";
