+++ libs/orpc/src/routers/app.router.ts
@@ -1,4 +1,4 @@
-import { heliusRouter } from "../helius/router";
+import { heliusRouter } from "./helius.router";
 
 export const appRouter = {
   helius: heliusRouter,
