+++ libs/orpc/src/routers/helius.router.ts
@@ -1,0 +1,5 @@
+import { getTokenList } from "../../procedures/getTokenList";
+
+export const heliusRouter = {
+  getTokenList,
+};
