--- libs/orpc/src/routers/helius/router.ts
@@ -1,5 +1,0 @@
-import { getTokenList } from "./procedures/getTokenList";
-
-export const heliusRouter = {
-  getTokenList,
-};
