+++ libs/orpc/src/rpcHandler.ts
@@ -1,8 +1,8 @@
 import { RPCHandler } from "@orpc/server/fetch";
 import { CORSPlugin } from "@orpc/server/plugins";
-import { appRouter } from "./routers/app/router";
+import { appRouter } from "./routers/app.router";
 
-export const handler = new RPCHandler(appRouter, {
+export const rpcHandler = new RPCHandler(appRouter, {
   plugins: [
     new CORSPlugin({
       allowMethods: ["GET", "HEAD", "PUT", "POST", "DELETE", "PATCH"],
